+++ src/utils.h
@@ -11,8 +11,8 @@ void shutdownScreen();
 void startRefresh();
 void endRefresh();
 
-void write(uint32_t row, uint32_t column, const char* str);
-void writeParsed(uint32_t row, uint32_t column, const char* str);
+void write(int row, int column, const char* str);
+void writeParsed(int row, int column, const char* str);
 
 void addToDownloadLog(char* str);
 void clearDownloadLog();
@@ -26,7 +26,7 @@ typedef enum {
 	B_RETURN__Y_RETRY = 2
 } ErrorOptions;
 
-void errorScreen(uint32_t line, ErrorOptions option);
+void errorScreen(int line, ErrorOptions option);
 void writeRetry();
 
 void enableShutdown();
@@ -34,6 +34,6 @@ void disableShutdown();
 
 char* b_tostring(bool b);
 
-char* hex(uint64_t i, uint8_t digits); //ex: 000050D1
+char* hex(uint64_t i, int digits); //ex: 000050D1
 
 #endif // ifndef WUPD_UTILS_H
